Work on improving both draw times and ability to control the presentation's appearance.
The main change in this commit is to implement a new ChartDC device context that wraps a wxDC with a flag to say whether antialiasing is available or not.

The idea is that the background is drawn using a plain old DC, which gives nice crisp gridlines and axis ticks, whilst any line plots use wxGCDC for nice smooth antialiasing. This also has the advantage of improving draw times because drawing the background on a plain old DC is much faster than on a wxGCDC.