Deprecate TextElementContainer(const wxString& title) ctor and introduce a new TextElementContainer(const TextElement& element) instead. Part of implementing enhancement #8.
This allows the font details to be specified as part of the ctor, which facilitates a more attractive default title.

Note: This commit only includes the TextElement and TextElementContainer classes (along with an updated wxfreechartdefs.h). This means that there will be deprecated warnings from those files that call the old ctor.