Update the way in which number axis intervals are calculated.
The old way was to set the total number of intervals (ticks), both major and minor, and then specify the number of minor ticks that constitute a major tick.

The new way is to define the *value* change for major intervals and then define the number of minor intervals per major interval.

Note: this has been done to pave the way for logarithmic scales.