Implement a very basic tooltip mechanism.
Currently it is only working with XY plots and vertical Bar plots (not stacked).

One of the limitations is the current way that Dataset stores data and the data that it stores. It would be better if all data was derived from a base class that has some meta information about that data.